refactor: migrate from classification to regression labeling
refactor: migrate from classification to regression labeling
feat: add regression loss (Huber) and ReduceLROnPlateau scheduler
refactor: update model output from 5 classes to continuous value
refactor: replace classification metrics with regression metrics (MAE, RMSE, R², Sign Accuracy)
feat: add regression visualizations (predictions, residuals, Q-Q plot)
docs: update README for regression approach
chore: update config.yaml for regression task
refactor: rename MambaClassifier to MambaRegressor, complete terminology migration
feat: implement clean minimal training logs with proper formatting
feat: update backtest for regression with continuous predictions and trading signals
docs: document trading principles in regression labeling strategy
fix: correct dtype from int64 to float32 for regression targets in sequence strategy
perf: optimize hyperparameters to prevent overfitting (smaller model, higher dropout)
fix: suppress torch.load and CUDA/TensorFlow warnings
chore: update ONNX export to use 'prediction' instead of 'logits'

+++ src/data/sequence_strategy.py
@@ -83,8 +83,8 @@ class SequenceProcessingStrategy(DataProcessingStrategy):
                 f"got {len(features)}"
             )
 
-        X = np.zeros((n_samples, self.sequence_length, features.shape[1]))
-        y = np.zeros(n_samples, dtype=np.int64)
+        X = np.zeros((n_samples, self.sequence_length, features.shape[1]), dtype=np.float32)
+        y = np.zeros(n_samples, dtype=np.float32)
 
         for i in range(n_samples):
             X[i] = features[i:i + self.sequence_length]
